Continue to work on Client side Token Ring
- Currently able to receive TOKEN from server and starts a
ClientSenderThread that is able to execute the event and then send the
token to the next client.

- Currently working on solving the issue of having clients connect to
each other - the issue with this is that each client can share a common
port # to make it easier to know where to connect to but the issue is
that we now cannot have more than 1 GUIclient since each client uses
port 4444 to listen…. This is bad.

— Once we figure this out, the ClientConnectionThread can be used to
create a connection and have an MSocket between the clients!