Successfully able to do a full connect
When a second client connects to the CNS, all clients are successfully
able to place each other on the map, and the newest client is now able
to properly setup a connection with the other clients.

The client that was listening for the connection now stores the mSocket
for that specific player for use later on.